Instrument for timing

+++ willump/graph/willump_graph_node.py
@@ -6,7 +6,7 @@ import willump.graph.utilities
 class WillumpGraphNode:
 
     def __init__(self, function_name: str, output_name: str, input_names: List[str],
-                 input_nodes: List['WillumpGraphNode'], cost: int, model_param: str = None):
+                 input_nodes: List['WillumpGraphNode'], cost: float, model_param: str = None):
         self.function_name = function_name
         self.output_name = output_name
         self.input_names = input_names
